Fix tree_corparch
1) incorrect nested dicts search
2) attach Manatee corpus info to CorpusInfo
