+++ lib/plugins/default_corparch/__init__.py
@@ -142,7 +142,7 @@ from lxml import etree
 import plugins
 from plugins.abstract.corpora import AbstractSearchableCorporaArchive
 from plugins.abstract.corpora import BrokenCorpusInfo
-from plugins.abstract.corpora import CorplistProvider, ManateeCorpusInfo
+from plugins.abstract.corpora import CorplistProvider, DefaultManateeCorpusInfo
 from plugins import inject
 import l10n
 import manatee
@@ -163,26 +163,6 @@ def translate_markup(s):
     return markdown(s.strip())
 
 
-class DcManateeCorpusInfo(ManateeCorpusInfo):
-    """
-    Represents a subset of corpus information
-    as provided by manatee.Corpus instance
-    """
-
-    def __init__(self, corpus, canonical_id):
-        super(DcManateeCorpusInfo, self).__init__()
-        self.encoding = corpus.get_conf('ENCODING')
-        import_string = partial(l10n.import_string, from_encoding=self.encoding)
-        self.name = import_string(corpus.get_conf('NAME') if corpus.get_conf('NAME')
-                                  else canonical_id)
-        self.description = import_string(corpus.get_info())
-        self.attrs = filter(lambda x: len(x) > 0, corpus.get_conf('ATTRLIST').split(','))
-        self.size = corpus.size()
-        attrlist = corpus.get_conf('ATTRLIST').split(',')
-        self.has_lemma = 'lempos' in attrlist or 'lemma' in attrlist
-        self.tagset_doc = import_string(corpus.get_conf('TAGSETDOC'))
-
-
 class ManateeCorpora(object):
     """
     A caching source of ManateeCorpusInfo instances.
@@ -194,13 +174,13 @@ class ManateeCorpora(object):
     def get_info(self, corpus_id):
         try:
             if corpus_id not in self._cache:
-                self._cache[corpus_id] = DcManateeCorpusInfo(
+                self._cache[corpus_id] = DefaultManateeCorpusInfo(
                     manatee.Corpus(corpus_id), corpus_id)
             return self._cache[corpus_id]
         except:
             # probably a misconfigured/missing corpus
-            return DcManateeCorpusInfo(EmptyCorpus(corpname=corpus_id),
-                                       corpus_id)
+            return DefaultManateeCorpusInfo(EmptyCorpus(corpname=corpus_id),
+                                            corpus_id)
 
 
 def parse_query(tag_prefix, query):
